Move ShaderNodeUtils to core
